Refactor helpers out to a separate file
